fix(auth-protection) - test bug

+++ src/components/auth-components/custom/useEffect/useLastRoutingEffect.tsx
@@ -7,8 +7,9 @@ import { routes } from '@/routing/router'
 export const useLastRoutingEffect = () => {
   const { asPath } = useRouter()
 
+  console.log(asPath)
   useEffect(() => {
-    if (asPath !== routes.unprotected) {
+    if (asPath !== routes.unprotected && undefined) {
       localStorage.removeItem('lastRouting')
       localStorage.setItem('lastRouting', asPath)
     }
